:recycle: [#6281] Clean up Objects API v2 registration code structure
Removed code duplication by refactoring into re-usable components
and/or hooks, and moved the code into a subfolder for the variables
table edit options.

This makes the UI components for the map, selectboxes and addressNL
components easier to read and maintain.

On top of that, the Formik state manipulation to manage the variable
mappings is simplified - we no longer need to deal with FieldArray
structures and can edit the options almost directly.
