+++ src/openforms/js/components/admin/form_design/registrations/objectsapi/edit_options/generic/index.js
@@ -1,0 +1,3 @@
+export {default as GenericEditor} from './GenericEditor';
+export {default as ShowJSONSchemaToggle} from './ShowJSONSchemaToggle';
+export {default as TargetPath} from './TargetPath';
