+++ src/openforms/js/components/admin/form_design/registrations/objectsapi/fields/DocumentTypes.js
@@ -2,7 +2,7 @@ import {useField, useFormikContext} from 'formik';
 import PropTypes from 'prop-types';
 import {useContext, useEffect, useMemo} from 'react';
 import {FormattedMessage, useIntl} from 'react-intl';
-import {useAsync, usePrevious} from 'react-use';
+import {usePrevious} from 'react-use';
 
 import {FeatureFlagsContext} from 'components/admin/form_design/Context';
 import Field from 'components/admin/forms/Field';
@@ -12,29 +12,12 @@ import {
   CatalogueSelect,
   CopyDocumentTypesConfig,
   DocumentTypeSelect as GenericDocumentTypeSelect,
-  getCatalogueOption,
-  groupAndSortCatalogueOptions,
-  useGetDocumentTypes,
 } from 'components/admin/forms/zgw';
 import {WarningIcon} from 'components/admin/icons';
-import {get} from 'utils/fetch';
 
-// Data fetching
+import {useGetDocumentTypes, useResolveCatalogue} from '../hooks';
 
-const CATALOGUES_ENDPOINT = '/api/v2/objects-api/catalogues';
-const IOT_ENDPOINT = '/api/v2/objects-api/document-types';
-
-const getCatalogues = async apiGroupID => {
-  const response = await get(CATALOGUES_ENDPOINT, {objects_api_group: apiGroupID});
-  if (!response.ok) {
-    throw new Error('Loading available catalogues failed');
-  }
-  return groupAndSortCatalogueOptions(response.data);
-};
-
-// Components
-
-const DocumentType = ({name, label, loading, documentTypes, isDisabled, helpText}) => {
+export const DocumentType = ({name, label, loading, documentTypes, isDisabled, helpText}) => {
   const intl = useIntl();
   const [{value}] = useField(name);
 
@@ -77,7 +60,7 @@ const DocumentType = ({name, label, loading, documentTypes, isDisabled, helpText
 };
 
 DocumentType.propTypes = {
-  name: PropTypes.oneOf(['iotSubmissionReport', 'iotSubmissionCsv', 'iotAttachment']).isRequired,
+  name: PropTypes.string.isRequired,
   label: PropTypes.node.isRequired,
   loading: PropTypes.bool.isRequired,
   isDisabled: PropTypes.bool.isRequired,
@@ -97,16 +80,11 @@ export const DocumentTypesFieldset = () => {
   // fetch available catalogues and re-use the result
   const {
     loading: loadingCatalogues,
-    value: catalogueOptionGroups = [],
+    catalogueOptionGroups,
     error: cataloguesError,
-  } = useAsync(async () => {
-    if (!objectsApiGroup) return [];
-    return await getCatalogues(objectsApiGroup);
-  }, [objectsApiGroup]);
+    catalogueUrl,
+  } = useResolveCatalogue(objectsApiGroup, catalogue);
   if (cataloguesError) throw cataloguesError;
-
-  const catalogueValue = getCatalogueOption(catalogueOptionGroups, catalogue || {});
-  const catalogueUrl = catalogueValue?.url;
   const previousCatalogueUrl = usePrevious(catalogueUrl);
 
   // if the catalogue changes, reset the selected document types
@@ -127,11 +105,7 @@ export const DocumentTypesFieldset = () => {
     }));
   }, [previousCatalogueUrl, catalogueUrl]);
 
-  const query = useMemo(
-    () => ({objects_api_group: objectsApiGroup, catalogue_url: catalogueUrl}),
-    [objectsApiGroup, catalogueUrl]
-  );
-  const {loading, documentTypes, error} = useGetDocumentTypes(IOT_ENDPOINT, query);
+  const {loading, documentTypes, error} = useGetDocumentTypes(objectsApiGroup, catalogueUrl);
   if (error) throw error;
 
   const documentTypeProps = {
